refactor: move test helper methods to separate test-only file
- Moved forceHashIndexEntry, getHashIndexCount, hasHashIndexEntry,
  and forceCleanupTrigger methods from types.go to test_helpers_test.go
- These methods are now excluded from production builds while remaining
  available for testing, keeping internals hidden and binary lean
- All tests continue to pass with helpers properly isolated
